Crate pest_ascii_tree

pest_ascii_tree

This is a small helper crate useful for quickly debugging your pest grammar. The rules found by parsing the file are formated into an ascii_tree.

It is useful, you you want to quickly debug your grammar without having to write specialized code for handling the Pairs iterator returned by your pest parser.

Example, for whan an output might look like.

 expr
 ├─ expr
 │  ├─ val "u"
 │  ├─ op "+"
 │  └─ expr
 │     ├─ val "v"
 │     ├─ op "+"
 │     └─ val "w"
 ├─ op "+"
 ├─ expr
 │  ├─ val "x"
 │  ├─ op "+"
 │  └─ val "y"
 ├─ op "+"
 └─ val "z"

Please, that the EOI rule is skipped.

Functions

into_ascii_tree
Formats the parsing result by pest into an ascii_tree into a String.

print_ascii_tree
Prints the result returned by your pest Parser as an ascii tree.